requirements/base.txt
python-dotenv == 1.0.0
Pillow == 9.4.0 
stomp.py == 8.1.0
pybase64 == 1.2.3
